requirements.txt
asgiref==3.5.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
Markdown==3.3.7
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.1.0
zipp==3.8.0
